// src/app/models/product.ts
export class Product {
    photoUrl: String;
    photoUrl2: String;
    photoUrl3: String;
    name: String;
    price: Number;
    descrition: String;
    disponibility: boolean;
    marca: String;
    modelo?: String;
    benefits?: String;
    treadwear?: String;
    traction?: String;
    temperature?: String;
    lowerCase: String;
    part?: String;
    vehicle?: String;
    categoria?: String;
}